pyproject.toml
[project]
name = "uart-mcp"
version = "2.0.0"
description = "MCP server for serial port communication with multi-session, full serial config, hex/base64 encoding, checksums, and project documentation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "pyserial>=3.5",
    "pyyaml>=6.0",
]

[project.scripts]
uart-mcp = "uart_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]